+++ packages/orq-rc/RELEASES.md
@@ -3069,3 +3069,13 @@ Based on:
 - [typescript v3.11.0-rc.17] packages/orq-rc
 ### Releases
 - [NPM v3.11.0-rc.17] https://www.npmjs.com/package/@orq-ai/node/v/3.11.0-rc.17 - packages/orq-rc
+
+## 2025-08-18 05:22:00
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.605.1 (2.684.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.11.0-rc.18] packages/orq-rc
+### Releases
+- [NPM v3.11.0-rc.18] https://www.npmjs.com/package/@orq-ai/node/v/3.11.0-rc.18 - packages/orq-rc